Revert "simplify access to "charm_shc" and "charm_pnmj""
While the new access to array elements in "charm_shc" and "charm_pnmj"
is more comfortable for user, I've realized it will be highly
impractical to use it when later parallelizing CHarm via MPI.  One would
either have to use the old access when compiled with MPI and the new
access when without MPI (highly inconvenient and confusing) or do not
allow direct access to spherical harmonic coefficients with the MPI
version of "charm_shc", which is also inconvenient.

This commit therefore reverts the commit that introduced the new access
to array elements.

This reverts commit 8f09bb87d9e22951bfdc48a3ab4875759277558b.