change the approach to "(R / r)^(n + 1)" in shs
In spherical harmonic synthesis, the "(R / r)^(n + 1)" terms are now
recursively recomputed on-the-fly over and over again which is in fact
faster then the single pre-computation outside the loops that was used
previously.  The speed improvement is probably because of better caching
and is most notable with high-degree syntheses.