improve performance of point sha and shs
This was achieved by improving CPU caching (cache blocking), dynamical
switching (added tiny and efficient loops to perform sha/shs exclusively
by F-numbers once X-numbers are no longer necessary) as well as by loop
unrolling.  For now, cache blocking and loop unrolling are not
implemented for sha and shs in cells.  To some extent, the dynamical
switching is applied with cells, though in a sub-optimal way, because no
F-numbers-only loops have been introduced yet.  All this might change in
the future if the performance of cell synthesis and analysis becomes an
issue.