add the first- and second-order gradients and guru
In CHarm, the following functions were added to the "shs" module:

    * charm_shs_point_grad1 (the full first-order gradient in LNOF),
    * charm_shs_point_grad2 (the full second-order gradient in LNOF),
    * charm_shs_point_guru (guru interface for first- and second-order
                            derivatives with respect to the spherical
                            coordinates).

In PyHarm, the new functions in the "pyharm.shs" module are:

    * pyharm.shs.point_grad1,
    * pyharm.shs.point_grad2,
    * pyharm.shs.point_guru.

Added new tests and reference test data for the new routines.

Added examples to cookbook on how to use the "grad1" and "grad2"
functions in CHarm and PyHarm.