define *_NMAX_* as symbolic constants instead of enum
"CHARM_SHC_NMAX_MODEL" and "CHARM_SHC_NMAX_ERROR" are now symbolic
constants instead of an enumeration.  This is because ISO C restricts
enumerator values to the range of "int", but we need the maximum values
of "unsigned long".  The latter values thus do not fall within the range
that can be portably stored by enumerations.