rename internal write function to array2file
It seems that "printf" on AArch64 (Android) uses some internal function
called "write".  The test suite defines a function called "write", too,
so there seems to be a collision, causing segfault when running the test
suite program.  I'm not sure, but I think this is not a problem on the
CHarm's side.  Anyway, to make it possible to execute the test suite
program on AArch64 (Android), too, the CHarm's "write" function was
renamed to "array2file".