define some integers in fnalfs computations as int64_t
The extended-range arithmetic method of Prof. Fukushima computes fnalfs
using integers as some auxiliary variables.  These were defined
previously as "int".  With NEON, they need to have the same bit width as
float or double, depending on the precision used to compile CHarm.  This
necessitated to define these integers as "int64_t".

More specifically, with Intel SIMD, we load the integers by explicitly
providing the values of the integers, which are then cast to "int64_t"
when calling the AVX-type intrinsics.  In this case, it is therefore OK
to define these integers as "int".

With NEON, however, we load the integers by referring to a pointer to an
array of integers.  This time, the array has to be therefore of the
"int64_t" type to ensure we load the proper amount of bits for each
integer.  Using "int" with the integers and NEON SIMD intrinsics would
be wrong.