remove NEG_R_INIT and ABS_R_INIT macros
They defined variables with bits properly set to swap the sign and
compute absolute value of AVX vectors.  Now, these variables are
embedded directly in NEG_R and ABS_R macros.  The rational is that such
variables are not needed with NEON.  Having them would result in the
compiler issuing lots of warnings.