improve checks of isfinite in configure.ac
The installation script is now able to detect if the "isfinite" macro
from "math.h" does not work correctly with the "-ffast-math" and similar
compiler options.  Previously, we checked for the presence of the macro
only and did not actually test it.  The test is important, because some
compiler options may cause the macro to not signalize nans.