prepare for a new release (v0.4.4)
Additional note to previous commit message: The issue was introduced in
v0.4.3, which required some modifications to "shs_point_kernel.c".  The
published PyHarm wheels for v0.4.3 are not affected by this issue, as
they rely on AVX and also "make check" during the build on GitHub passed
all tests.