fix compiling errors for old versions of gcc
Some old versions of gcc choke when variables are declared with the
"const" qualifier and enter OpenMP block as "shared".  The "const"
qualifiers were therefore removed.  Alternatively, we could not label
them as "shared" instead, but this causes another error with newer
compilers, as we use also the "default(none)" option.